Separate physics package from detector coatings
